+++ CHANGELOG.md
@@ -5,10 +5,15 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [1.5.0] - 2019-01-27
 ### Added
- - CSS classes to DataTableHelper.
- - Methods to StringHelper class.
- - Methods to ListHelper class.
+- CSS classes to DataTableHelper.
+- Methods to StringHelper class.
+- Methods to ListHelper class.
+
+### Removed
+- Moved classes to ADN.Extensions and ADN.IO
 
 ## [1.4.0] - 2019-01-12
 ### Added